+++ src/locales/nl/queue.ts
@@ -13,7 +13,7 @@ export const queue = {
   appendToQueue: 'Toevoegen aan wachtrij',
   delete: 'Verwijderen',
   deleteConfirm: 'Afspeellijst "{{name}}" verwijderen?',
-  clear: 'Leegmaken',
+  clear: 'Wachtrij leegmaken',
   shuffle: 'Wachtrij shufflen',
   gapless: 'Naadloos',
   crossfade: 'Overgang',